refactor(server): split admin.routes into feature modules
- admin-stats-and-health: cleanup, DB stats, consolidation telemetry, stubs
- admin-memory-review: item preview, review queue, SSE, review/dismiss
- admin-batch: scheduler status/history/run, sleep consolidation
- admin-runtime-performance: PerformanceMonitor metrics/alerts/summary
- admin-tools: anchor restore, embedding migrate, episodic conversion, meta-stats
- admin-project-memory: project stats and cleanup (Issue #81)
- admin.routes.ts: compose register* calls only; export graph types unchanged

Relates-to: #315
